Restructured class system
Styled the webpage using the div element mostly.

Removed the align property in the image elements and used the float attribute combined with two classes for aligning the image the right/left.

Reset padding and margins for all elemnts to 0.

Centered the site and gave it a maximum and minimum width.

Now a managable structure for adding blog entries has been set.